requirements.txt
cffi==1.17.1
cryptography==44.0.2
cx-oracle==8.3.0
demographics==0.0.1
greenlet==3.1.1
mysql-connector-python==9.2.0
numpy==2.0.2
oracledb==3.0.0
pandas==2.2.3
pycparser==2.22
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlalchemy==2.0.39
typing-extensions==4.12.2
tzdata==2025.2
